drvloc/redis: Add & implement Close method
Add and implement the Close method for being able to release the
resources and close the Redis connection hold by an instance of the
service.

Due to this new method and what it involves in terms of controlling when
the service is closed or not, a refactoring has been done in the
constructor function (for returning the specific type) and in the
methods in order of returning a specific error in case that they are
called after the Close method has been called.

On the refactoring a error returned by the LocationsForLastMinutes
method was spotted of not being very precise, so a new error code has
been added for it and replaced in such place.