zmbdrv/drvloc: Define Service & impl ctor func
Define the type which will implement the Zombie Driver Service which
relies on the functionalities exposed by the Driver Location Service
using a client of it.

A new entity has also been added for making configurable the set of
rules which determines if a driver is a zombie or not.