zmbdrv/drvloc: Impl IsZombie method
Implement the IsZombie method of the Zombie Driver Service which relies
on the functionality of the Driver Location Service and its test.

Because some test helper functions which were in an internal package of
the drvloc package were needed for the IsZombie test, such internal
package has been moved to the root of the repository in order of being
accessible by all the packages of this repository and avoiding to
duplicate code between them.