gateway: Define & impl endpoints & ctor
Define and implement the type which holds the HTTP handlers exposed on
the API endpoints and a constructor function which instance a new HTTP
server and configure the endpoints.

The endpoints routing and the handler logic is tested with an
integration test, for guaranteeing that the routes are properly set
based on the configuration and the handlers use the Driver Location and
Zombie Driver services properly.

Because it isn't wanted to pass HTTP configurations to gateway
constructor function because it only uses the Driver Location Service
functionalities exposed over NSQ, a new client NSQHTTP Driver Location
Service client has been created for exposing only the functionalities
which are served over NSQ.

Ideally, there should be 2 transports of the Driver Location Service,
one for exposing fully the service over HTTP and another for doing it
over NSQ and the one which is a mix, should be as now, but using the
needed parts of the both implementations, but that's out of the context
of this project for now.