zmbdrv/http/cmd/server: Impl main
Implement a main package for running the Zombie Driver Service behinds
a HTTP server.

Because it isn't wanted to pass NSQ configurations to this main and the
Zombie Driver Service only uses the Driver Location Service
functionalities exposed over HTTP, a new client NSQHTTP Driver Location
Service client has been created for exposing only the functionalities
which are served over HTTP.

Ideally, there should be 2 transports of the Driver Location Service,
one for exposing fully the service over HTTP and another for doing it
over NSQ and the one which is a mix, should be as now, but using the
needed parts of the both implementations, but that's out of the context
of this project for now.