Makefile: Add e2e target & update clean
Add a target to the general Makefile to run some request to the default
configuration of the services to realize that the whole system works.

The clean target has been updated to call the clean target of each
subpackage Makefile.

The default config.yaml file has been updated for being usable with as
its with the simple end to end test run by the e2e Makefile target.